Add script to create updated RPI image
update_pi_img.sh is a utility script that can build a RPI image based on
an older image file. For example, the v1.2 zip file can be used to
generate the v1.3 img automatically.

The script effectively opens the image, runs a 'git pull' on it and
saves it as a new version.

NOTE/TODO: most of the time is spent extracting / recompressing the zip
file. The actual script itself is much faster...